Use deployContracts function in integration tests
